Extract message into floating point arithmetic; non-floating point option

+++ src/com/pavdem2012/helloworld/HelloWorld.java
@@ -3,9 +3,9 @@ package com.pavdem2012.helloworld;
 public class HelloWorld {
 
 	public static void main(String[] args) {
-		int i = 6; 
-		int j = 7; 
-		float answer =(float)i /(float)j; 				
+		float i = 6; 
+		float j = 7; 
+		float answer =i /j; 				
 	String message = "Hello World!!!! Да ну нафиг так бывает???";
 	String stringAnswer = message + " " + answer;
          System.out.println(message);
